refactor: extract PlaceAtDistanceContainer
